requirements.txt
appdirs==1.4.3
asgiref==3.2.10
backcall==0.2.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
Django==3.1
django-environ==0.4.5
django-extensions==3.0.4
django-material-admin==1.7.8
djangorestframework==3.11.1
et-xmlfile==1.0.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
ipython==7.17.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.17.2
lockfile==0.12.2
msgpack==0.6.2
openpyxl==3.0.4
packaging==20.3
parso==0.7.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
progress==1.5
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
traitlets==4.3.3
urllib3==1.25.8
uWSGI==2.0.19.1
wcwidth==0.2.5
webencodings==0.5.1
